+++ fxmetrics/README.md
@@ -5,6 +5,7 @@ This module provides [prometheus metrics](https://pkg.go.dev/github.com/promethe
 This package provides 2 modules:
 
 * A regular module that exposes a standard prometheus endpoint for use in long running daemons
+* An OTLP module which uses the prometheus sdk to define metrics, but pushes them out over OTLP rather than the standard prometheus http endpoint
 * A push module that pushes metrics into push gateway, for use with jobs
 
 Because client code only interacts with the `*prometheus.Registry`, they can be swapped out transparently.
@@ -27,11 +28,25 @@ The module lazily provides the following components:
 
 It starts an additional webserver exposing the prometheus endpoint.
 
+## OTLP Module
+
+### Components 
+The module lazily provides the following components:
+
+* A `*prometheus.Registry`
+* A `metric.MeterProvider` (allows you to define metrics with the otel sdk)
+* GrpcServerInterceptors that count all incoming requests by method and status
+* GrpcClientInterceptors that count all requests made with the client by method and status
+
+It adds hooks to push the metrics when the system stops and at regular intervals during runtime.
+
 ### Configuration
 The module provides the following configuration options:
-* `Server`: A http server config, see the docs in the fxhttp package for details
+* `GrpcClient`: A grpc client config, see the docs in the fxgrpc package for details
 * `Histograms`: A bool which enables support for histograms in the grpc middleware (will most likely be removed)
 * `ProcessName`: A string used as a prefix inside the process collector to prevent clashes
+* `PushInterval`: The frequency at which metrics are pushed during runtime
+* `Enabled`: Disables the pushing of metrics completely
 
 ## Push Module
